chore(cloud): clean managed sandbox repo shape
