+++ scripts/server_boundaries_allowlist.txt
@@ -30,9 +30,10 @@ SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/live/service.py 4 remaini
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/worker/transactions.py 2 Lane 7 explicit worker API pre-response commit boundary debt
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/lifecycle/service.py 12 lifecycle extraction preserves existing workspace lifecycle transaction/session boundary debt through session_ops
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/provisioning/preflight.py 6 provisioning preflight extraction preserves existing repo/config validation transaction/session boundary debt through session_ops
-SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/provisioning/service.py 7 provisioning extraction preserves existing workspace provisioning transaction/session boundary debt through session_ops
+SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/provisioning/service.py 12 managed sandbox finalization preserves existing workspace provisioning transaction/session boundary debt through session_ops
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/remote_access/service.py 4 remote-access connection flow preserves existing explicit session boundary debt through session_ops
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/target_launch/service.py 7 target-launch orchestration preserves existing explicit command/session boundary debt through session_ops
 DOMAIN_FORBIDDEN_IMPORT server/proliferate/server/cloud/agent_run_config/domain/resolve.py 1 transitional agent run config resolver imports product constants
 DOMAIN_FORBIDDEN_IMPORT server/proliferate/server/cloud/target_config/domain/policy.py 1 transitional target config policy imports product config during worker migration
 DOMAIN_FORBIDDEN_IMPORT server/proliferate/server/cloud/targets/domain/policy.py 1 transitional target policy imports product config during worker migration
+PRODUCT_RAW_HTTP_IMPORT server/proliferate/server/cloud/gateway/proxy.py 1 managed sandbox gateway owns temporary HTTP/WS proxy transport until gateway access moves below product service layer
